#3 set uint64 to fixed width type

+++ ASMC_SRC/SRC/Types.hpp
@@ -17,10 +17,13 @@
 #ifndef TYPES_HPP
 #define TYPES_HPP
 
+#include <cstdint>
+#include <cinttypes>
+
 typedef unsigned char uchar;
 typedef unsigned int uint;
-typedef unsigned long long uint64;
-typedef long long int64;
+typedef uint64_t uint64;
+typedef int64_t int64;
 
 struct uint64_masks {
 	uint64 is0, is2, is9;
